requirements.txt
# Core dependencies
flask>=2.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.21.0

# Document processing
pdfplumber>=0.9.0
python-docx>=0.8.11
PyYAML>=6.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# Web server (production)
waitress>=2.1.0
gunicorn>=21.0.0

# Optional OCR support
# ocrmypdf>=15.0.0  # Uncomment if you need OCR

# Development dependencies (optional)
# black>=23.0.0
# pytest>=7.4.0
# pytest-cov>=4.1.0 